// frontend/src/app/shared/generated-api/model/notebookView.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentColor } from './contentColor';


export interface NotebookView { 
    id: string;
    name: string;
    color: ContentColor;
    defaultNotebook: boolean;
    sortOrder: number;
    noteCount: number;
    createdAt: string;
    updatedAt: string;
    version: number;
}